+++ Models/LiveApiKey.cs
@@ -1,0 +1,10 @@
+namespace Facturapi
+{
+    public class LiveApiKey
+    {
+        public string First12 { get; set; }
+        public string CreatedAt { get; set; }
+
+        public string Id { get; set; }
+    }
+}