+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.6.0] - 2024-23-09
+
+### Added
+
+- Add List of Live Api Keys `Organizations.ListAsyncLiveApiKey`
+- Add Delete of a Live Api Key `Organization.DeleteAsyncLiveApiKey`
+
 ## [4.5.0] - 2024-06-05
 
 ### Added
@@ -32,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update invoice uses constants
 - Update payment form constants
 - Add new method for delete certs in organization
+
 
 ## [4.1.0] - 2023-12-06